Fixed healthbar color

+++ coolgame/HealthBar.cs
@@ -25,7 +25,7 @@ namespace coolgame
                 val = value;
                 rectangle.Width = (int)(100 * val);
                 rectangle.Height = 10;
-                color = new Color(Math.Min(255, 500 * (1 - val)), Math.Min(255, 500 * val), 0);
+                color = new Color(Math.Min(255, (int)(500 * (1 - val))), Math.Min(255, (int)(500 * val)), 0);
             }
         }
